blockstore: use `alloc` for block buffer allocation for now
There is potential for around a 10% to 23% increase to the performance
of block reads by using `allocUnsafe`, however there is already around
a 3 to 6 times increase to the performance, and this optimization can
be added later. While it's safe to use `allocUnsafe` as the number of
bytes read is checked to be the same size as the buffer allocation,
there is a potential for test cases to introduce _other_ behavior
for `fs.read` that may not have the same behavior, though this
isn't currently the case.